refactor(emitter-go): split rendering concerns into abstractions
