+++ packages/emitter-go/src/render/index.ts
@@ -1,0 +1,8 @@
+export { renderDiagnosticsComments } from "./diagnostics-comment-rendering";
+export {
+  renderGoImports,
+  renderMainFunction,
+  renderResolveListenAddr,
+  renderRouteRegistry,
+} from "./server-bootstrap-rendering";
+export { renderRoute } from "./template-rendering";
